UI improvements made, password change feature implemented and code refactored.

+++ server/controllers/user.js
@@ -2,6 +2,7 @@ const Users = require("../models/user");
 const _ = require("lodash");
 const path = require("path");
 const fs = require("fs");
+const bcrypt = require("bcrypt");
 
 exports.getUsers = async(req,res, next)=>{
     try{
@@ -35,35 +36,63 @@ exports.getUser = async(req,res, next)=>{
 
 exports.updateUser = async (req,res, next)=>{
     try{
-        const reqBody = JSON.parse(req.body.user);
+        const edit = req.query.edit;
         const user = await Users.findById(req.params.userId);
+        let reqBody;
         if(req.userId == user._id){
-            if(user.email!=reqBody.email){
-                const doesUserExist = await Users.find({email: reqBody.email});
-                if(doesUserExist.length>0){
-                    const error = new Error("User already exists!");
-                    error.statusCode = 409;
-                    error.data = "User already exists!";
+            switch(edit){
+                case "picture":
+                    let profilePicture = user.profilePicture;
+                    if(req.file){
+                        profilePicture = "/"+req.file.path.replace("\\", "/");
+                    }
+                    if(profilePicture!=user.profilePicture && user.profilePicture!="/images/default.png"){
+                        clearImage(user.profilePicture);
+                    }
+                    user.profilePicture = profilePicture;
+                    break;
+                case "email":
+                    reqBody = JSON.parse(req.body.user);
+                    if(user.email!=reqBody.email){
+                        const doesUserExist = await Users.find({email: reqBody.email});
+                        if(doesUserExist.length>0){
+                            const error = new Error("User already exists!");
+                            error.statusCode = 409;
+                            error.data = "User already exists!";
+                            throw error;
+                        }
+                    }
+                    _.extend(user, {...reqBody});
+                    break;  
+                case "username":
+                    reqBody = JSON.parse(req.body.user);
+                    _.extend(user, {...reqBody});
+                    break;
+                case "password":
+                    reqBody = JSON.parse(req.body.user);
+                    const passwordMatches = await bcrypt.compare(reqBody.oldPassword, user.password);
+                    if(passwordMatches){
+                        const encryptedPassword = await bcrypt.hash(reqBody.newPassword, +process.env.SALT_ROUNDS);
+                        user.password=encryptedPassword;
+                    }else{
+                        const error = new Error("Old password incorrect!");
+                        error.statusCode = 401;
+                        error.data = "Old password incorrect!";
+                        throw error;
+                    }
+                    break;
+                default:
+                    const error = new Error("No edit specified.");
+                    error.statusCode = 404;
+                    error.data = "No edit specified.";
                     throw error;
-                }
-            }
-            let profilePicture = reqBody.profilePicture;
-            if(req.file){
-                profilePicture = "/"+req.file.path.replace("\\", "/");
-            }
-            if(profilePicture!="" && profilePicture!=undefined && profilePicture!=null && profilePicture!=user.profilePicture && user.profilePicture!="/images/default.png"){
-                clearImage(user.profilePicture);
-            }
-            _.extend(user, {...reqBody, profilePicture: user.profilePicture});
-            if(profilePicture!="" && profilePicture!=undefined && profilePicture!=null){
-                user.profilePicture = profilePicture;
             }
             user.updatedAt = Date.now();
             const updatedUser = await user.save();
             updatedUser.password = undefined;
             return res.json({
                 user: updatedUser,
-                message: "User successfully updated!"
+                message: "Picture successfully updated!"
             })
         }else{
             const error = new Error("Unauthorized!");
@@ -79,17 +108,36 @@ exports.updateUser = async (req,res, next)=>{
     }
 }
 
+exports.updatePassword = async (req,res,next)=>{
+    try{
+
+    }catch(error){
+        if(!error.statusCode){
+            error.statusCode=500;
+        }
+        next(error);
+    }
+}
+
 exports.deleteUser = async (req,res,next)=>{
     try{
         if(req.userId == req.params.userId){  
             const user = await Users.findById(req.userId);
-            if(user.profilePicture && user.profilePicture!="/images/default.png"){
-                clearImage(user.profilePicture);
+            const passwordMatches = await bcrypt.compare(req.body.password, user.password);
+            if(passwordMatches){
+                if(user.profilePicture && user.profilePicture!="/images/default.png"){
+                    clearImage(user.profilePicture);
+                }
+                await Users.findByIdAndDelete(req.userId)
+                return res.json({
+                    message: "User successfully deleted!"
+                })
+            }else{
+                const error = new Error("Password incorrect!");
+                error.statusCode = 401;
+                error.data = "Password incorrect!";
+                throw error;
             }
-            await Users.findByIdAndDelete(req.userId)
-            return res.json({
-                message: "User successfully deleted!"
-            })
         }else{
             const error = new Error("Unauthorized!");
             error.statusCode = 401;
